requirements.txt
tensorflow==2.15.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.1
matplotlib==3.8.4
opencv-python==4.10.0.84
Pillow==10.4.0
pydicom==2.4.4
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
starlette==0.38.2
albumentations==1.4.10